requirements.txt
python-dotenv~=1.0.0
python-telegram-bot[webhooks]~=20.2
pymongo~=4.3.3
requests~=2.28.2
Pillow~=9.5.0